R6.1: split vector/state into model/build/check/simulation
Facade mod.rs re-exports everything; no public path changes. Producer
construction (build_vector_state_plan{,_with_clock,_with_resources} and
producer-only helpers effective_delay_requirements,
promote_lockstep_register_delays) lives in build.rs; the standalone
checker entries and the shared terminal verify path live in check.rs;
bounded-simulation state machines in simulation.rs; vocabulary + error
taxonomy in model.rs.

Plan §4.8 — admission guards in the shared verify function
`verify_vector_state_plan_after_vector_plan` (hosted in check.rs):
- verify_source_alignment
- verify_supported_state
- verify_delays (incl. independently_expected_delay_requirements)
- verify_recursions
- verify_phases
- expected_no_op_resources / expected_special_transitions coverage

remain on BOTH paths after the split: the producer's terminal
verification (build.rs) and the standalone checker entries
(verify_vector_state_plan{,_with_clock,_with_resources}). Existing
checker-entry rejection tests
(independent_checker_rejects_geometry_and_phase_mutations,
lockstep_delay_one_register_mapping_is_checked_independently) still
fail through the checker entry alone and stay green.

Validation: transform 387 lib tests, workspace clippy (+1.97.0) clean.
