+++ crates/transform/src/signal_fir/vector/state/tests.rs
@@ -1,16 +1,26 @@
 //! Tests for `vector::state` (relocated from the former inline
 //! `mod tests` block; test names unchanged).
 
+use std::collections::BTreeMap;
+
 use propagate::ClockDomainTable;
 use signals::SigBuilder;
 use tlib::TreeArena;
 
+use super::build::{build_vector_state_plan_with_resources, effective_delay_requirements};
+use super::check::verify_vector_state_plan_with_resources;
 use super::*;
 use crate::clk_env::annotate;
+use crate::signal_fir::decoration_verify::VerifiedDecorationCertificate;
 use crate::signal_fir::decoration_verify::certify_decorations;
 use crate::signal_fir::pv_slice::build_pv_signals;
+use crate::signal_fir::vector::analysis::DepKind;
 use crate::signal_fir::vector::clock_ad::build_vector_clock_ad_plan;
+use crate::signal_fir::vector::plan::VerifiedVectorPlan;
 use crate::signal_fir::vector::plan::{build_vector_plan, build_vector_plan_with_lockstep};
+use crate::signal_fir::vector::verify::LoopKind;
+use crate::signal_fir::vector::verify::Placement;
+use crate::signal_prepare::VerifiedPreparedSignals;
 use crate::signal_prepare::prepare_signals_for_fir_verified;
 
 fn certify(arena: &TreeArena, roots: &[signals::SigId]) -> VerifiedDecorationCertificate {
