+++ porting/journal/2026-07-20.md
@@ -49,3 +49,26 @@ modules on both sides of the plan artifact. Validation: transform 385,
   vérification terminale du producteur.
 - Gates : transform 387 (385+2), vector_mode 35/35, clippy workspace
   +1.97.0 vert, fmt vert. Arbitre en cours après commit.
+
+## R6.1 — scission de `vector/state`
+
+- `state/mod.rs` (2 017 lignes) scindé en `model.rs` (399), `build.rs` (356),
+  `check.rs` (1 020), `simulation.rs` (203), façade `mod.rs` (35).
+- Correctif parseur : les blocs `impl<T: Clone>` génériques échappaient à
+  l'inventaire (regex exigeait `impl ` avec espace) — les deux impl de
+  simulation (CopyDelayState/RingDelayState) repris par plage de lignes.
+- Classement par graphe d'appels : helpers producteur-seuls
+  (`effective_delay_requirements`, `promote_lockstep_register_delays`,
+  `build_vector_state_plan_with_resources`) dans `build.rs` ; tout le
+  vocabulaire partagé et checker (delay_storage, canonical_lockstep_…,
+  expected_*, verify_*) dans `check.rs`.
+- §4.8 : `verify_vector_state_plan_after_vector_plan` (hébergeant
+  verify_source_alignment, verify_supported_state, verify_delays,
+  verify_recursions, verify_phases et les comparaisons expected_no_op /
+  expected_special) reste dans `check.rs`, appelé par la vérification
+  terminale du producteur ET par les entrées checker autonomes. Tests de
+  rejet par l'entrée checker déjà présents
+  (`independent_checker_rejects_geometry_and_phase_mutations`,
+  `lockstep_delay_one_register_mapping_is_checked_independently`) —
+  inchangés et verts.
+- Gates : transform 387, clippy workspace vert, fmt vert.
